CRC32 patch

+++ SRC/crc32.h
@@ -1,7 +1,8 @@
 #ifndef __CRC32_H
 #define __CRC32_H
+#include <stdint.h>
 
 unsigned long crc32_buf(const void *buf, long len);
-unsigned long crc32_file(const char *filename);
+uint32_t  crc32_file(const char *filename);
 
 #endif
